+++ client/__tests__/components/write/PostEditor.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent, render, RenderResult } from '@testing-library/react';
-import PostEditor, { PostEditorProps } from '@/components/write/PostEditor';
+import PostEditor, { PostEditorProps } from '@/components/write/Editor/PostEditor';
 
 const DEFAULT_ARGS: PostEditorProps = {
   onChange: jest.fn(),
