+++ client/src/components/write/Modal/WriteModal.styles.tsx
